# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/timesheets": {
            "get": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Get all timesheets",
                "operationId": "7d14e54a11044a9563dc96a368456248",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Timesheet"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Create a new timesheet",
                "operationId": "8ba9cc447e3f3bec76dcea5d1b31250e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "task_name": {
                                        "description": "Name of the task",
                                        "type": "string",
                                        "example": "Task 1"
                                    },
                                    "date": {
                                        "description": "Date of the timesheet entry",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-02-20"
                                    },
                                    "hours": {
                                        "description": "Number of hours worked",
                                        "type": "number",
                                        "format": "float",
                                        "example": 8
                                    },
                                    "user_id": {
                                        "description": "ID of the user who worked on the task",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "project_id": {
                                        "description": "ID of the project the task belongs to",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Timesheet created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Timesheet"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/timesheets/{timesheet}": {
            "get": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Get a single timesheet by ID",
                "operationId": "e1f73033fe98f8c89f323ddd20dd27c4",
                "parameters": [
                    {
                        "name": "timesheet",
                        "in": "path",
                        "description": "Timesheet ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Timesheet"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Timesheet not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Update an existing timesheet",
                "operationId": "658d246fd5b31661bb6ce5ee47466731",
                "parameters": [
                    {
                        "name": "timesheet",
                        "in": "path",
                        "description": "Timesheet ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "task_name": {
                                        "description": "Name of the task",
                                        "type": "string",
                                        "example": "Task 1",
                                        "nullable": true
                                    },
                                    "date": {
                                        "description": "Date of the timesheet entry",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-02-20",
                                        "nullable": true
                                    },
                                    "hours": {
                                        "description": "Number of hours worked",
                                        "type": "number",
                                        "format": "float",
                                        "example": 8,
                                        "nullable": true
                                    },
                                    "user_id": {
                                        "description": "ID of the user who worked on the task",
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "project_id": {
                                        "description": "ID of the project the task belongs to",
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Timesheet updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Timesheet"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Delete a timesheet",
                "operationId": "f765a7ec3bce635ac0487d9079e6beb5",
                "parameters": [
                    {
                        "name": "timesheet",
                        "in": "path",
                        "description": "Timesheet ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Timesheet deleted successfully"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Timesheet": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "task_name": {
                        "type": "string",
                        "example": "Develop API"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-02-20"
                    },
                    "hours": {
                        "type": "number",
                        "format": "float",
                        "example": 8.5
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "project_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}